initial: add giveaway and simple site generator
This uses "build.py" to render a tree of rst to html, with no templating.

I considered using an existing generator, but most are way too heavyweight for
what we need and what I'm willing to learn right now.

bluespan-normalize.css is currently static, but I expect to use some form of css
composition/generation eventually.